Added logging

+++ app/book/loan_books.py
@@ -1,4 +1,5 @@
 import datetime
+import logging
 from datetime import date
 
 from fastapi import APIRouter, Depends, HTTPException, status
@@ -10,6 +11,7 @@ from app.user.auth import UserResponse, get_current_user
 from app.book.books import BookResponse
 
 router = APIRouter()
+logging.basicConfig(filename='app.log', level=logging.INFO, format='%(asctime)s - %(levelname)s - %(module)s - %(message)s')
 
 
 class LoanCreate(BaseModel):
@@ -50,6 +52,7 @@ def take_book(book_id: int, current_user: User = Depends(get_current_user), db: 
     db.commit()
     db.refresh(db_loan)
     db.refresh(db_book)
+    logging.info(f"{current_user.username} take the book with ID: {book_id}")
     return db_loan
 
 
@@ -61,11 +64,14 @@ def return_book(book_id: int, current_user: User = Depends(get_current_user), db
     db_loan = db.query(Loan).filter(Loan.user_id == current_user.id,
                                     Loan.book_id == book_id
                                     ).first()
+    if not db_loan:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Not found loans by user or book")
     db.delete(db_loan)
     db_book = db.query(Book).filter(Book.id == db_loan.book_id).first()
     db_book.copies = db_book.copies + 1
     db.commit()
     db.refresh(db_book)
+    logging.info(f"{current_user.username} return the book with ID: {book_id}")
     return {"detail": "Returned book", "book_id": book_id, "user_id": current_user.id}
